requirements.txt
camelot-py[cv]
pandas
beautifulsoup4